test: update user routes
